feat: refactor module exports and configurations across files
- Update the `libs/src/index.ts` file with new exports and remove some existing exports
- Add a new `libs/tailwind.config.js` file with custom theme and extension configurations
